+++ examples/api/brunel.py
@@ -15,7 +15,7 @@ import matplotlib
 from pyNN.utility import SimulationProgressBar
 import sys
 argv = sys.argv[1:]
-from pype9.mpi import is_mpi_master  # @IgnorePep8
+from pype9.utils.mpi import is_mpi_master  # @IgnorePep8
 import pype9.utils.logging.handlers.sysout  # @UnusedImport @IgnorePep8
